김일훈

+++ JavaJourney/src/main/java/web/dto/Bean.java
@@ -28,7 +28,7 @@ public class Bean {
 	// beanRev join
 	private int revStarPoint;
 	private int reviewCount;
-	private int avgRevStarPoint;
+	private double avgRevStarPoint;
 	
 	// beanSub join
 	private int gram;
